long constant instruction
Allows for handling of more then 255 constant values in a code chunk.

+++ value.c
@@ -11,7 +11,7 @@ void initValueArray(ValueArray *array) {
 
 void writeValueArray(ValueArray *array, Value value) {
     if (array->capacity < array->count + 1) {
-        uint8_t oldCapacity = array->capacity;
+        size_t oldCapacity = array->capacity;
         array->capacity = GROW_CAPACITY(oldCapacity);
         array->values = GROW_ARRAY(Value, array->values, oldCapacity, array->capacity);
     }
